% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdefrost.R
\name{ggdefrost}
\alias{ggdefrost}
\title{This function is used to produce a single heatmap which combines two different correlation coefficients (Pearson's R and Spearman's Rho). Variables will be ordered from left to right and bottom to top as identified in 'vars1' and' vars2'.}
\arguments{
\item{data}{A data frame consisting of the variables to be included in the heatmap.}

\item{vars1}{The first of two possible list of variables to be included in the heatmap.}

\item{vars2}{The second of two possible list of variables to be included in the heatmap.}

\item{type1}{The type of correlation coefficient to calculate for the variables listed in the 'vars1' argument. One of c("pearson","spearman"), with the default being "pearson".}

\item{type2}{The type of correlation coefficient to calculate for the variables listed in the 'vars2' argument. One of c("pearson","spearman"), with the default being "spearman".}

\item{diff}{The characteristic used to differentiate the variables listed in the 'vars2' argument. One of c("outline","round"), with the default being "outline".}

\item{alpha}{The alpha value to determine significance, in which significant p-values will have the corresponding correlation coefficient in the heatmap denoted in bold text. The default alpha value is 0.05.}

\item{coef_size}{The size of the correlation coefficient in the heatmap. The default is 7.}

\item{colour}{The colour of the outline for variables identified in 'vars2' when 'diff' is set to "outline". The default is a dark grey.}

\item{triangle}{Whether the heatmap takes on a square or triangle form. The default is FALSE.}

\item{vars_rename}{Optional names to reassign to the variables (in order of the variables identified in the vars1 and vars2 arguments).}
}
\value{
A ggplot graphical object.
}
\description{
This function is used to produce a single heatmap which combines two different correlation coefficients (Pearson's R and Spearman's Rho). Variables will be ordered from left to right and bottom to top as identified in 'vars1' and' vars2'.
}
\examples{

### --- Load Simulated Data

data("reheat_data")

### --- Heatmap with Pearson and Spearman correlations

### The default 'type1' argument is Pearson correlations for 'vars1' variables
### The default 'type2' argument is Spearman correlations for 'vars2' variables
### Variables in 'vars2' argument will be denoted with a round tile

ggdefrost(reheat_data,
vars1 = c("audit_score","sip_score","drinks_per_week","cudit_score","ftnd_score"),
vars2 = c("cannabis_assist","cigarette_assist","ecigarette_assist"),
diff = "round")

}
