% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggreheat.R
\name{ggreheat}
\alias{ggreheat}
\title{This function is used to produce a heatmap using a single correlation coefficient (Pearson's r or Spearman's rho). It can also create side-by-side heatmaps for
a single binary grouping variable. When calculating Pearson's r, confidence intervals produced from Fisher Z transformations can also be displayed alongside the
correlation coefficient. If creating side-by-side heatmaps, the confidence intervals which do not overlap between the two groups will be highlighted.}
\arguments{
\item{data}{A data frame consisting only of the variables to be included in the heatmap.}

\item{vars}{The list of variables to be included in the heatmap.}

\item{type}{The type of correlation coefficient to calculate. The default is "pearson". One of c("pearson","spearman").}

\item{by}{A binary variable to create side-by-side heatmaps. If defined, side-by-side heatmaps will be created by the binary variable.}

\item{alpha}{The alpha value to determine significance, in which significant p-values will have the corresponding correlation coefficient in the heatmap denoted in bold text. The default alpha value is 0.05.}

\item{include_ci}{Whether confidence intervals calculated using Fisher Z transformations should be displayed in the heatmap. Only available when 'type' is "pearson". The default is TRUE.}

\item{coef_size}{The size of the correlation coefficient in the heatmap. The default is 7.}

\item{colour_ci}{The colour of the outline for highlighting Fisher Z transformations CIs that do not overlap. Available when type is "pearson" and a 'by' variable is defined. The default is a dark grey.}

\item{tile_type}{The type of tile corner to use in the heatmap. One of c("square","round"), with the default being "square".}

\item{triangle}{Whether the heatmap takes on a square or triangle form. The default is FALSE.}

\item{vars_rename}{Optional names to reassign to the variables listed in the vars argument.}
}
\description{
This function is used to produce a heatmap using a single correlation coefficient (Pearson's r or Spearman's rho). It can also create side-by-side heatmaps for
a single binary grouping variable. When calculating Pearson's r, confidence intervals produced from Fisher Z transformations can also be displayed alongside the
correlation coefficient. If creating side-by-side heatmaps, the confidence intervals which do not overlap between the two groups will be highlighted.
}
