% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reheat_data.R
\docType{data}
\name{reheat_data}
\alias{reheat_data}
\title{An example data set from simulated data. The variables are as follows: "id" = participant unique identifier; "group" = group participant belongs to (A or B);
"audit_score" = Alcohol Use Disorders Identification Test (AUDIT) total score; "sip_score" = Short Inventory of Problems (SIP) total score;
"drinks_per_week" = total number of standard drinks endorsed in a typical week; "cannabis_assist" = Alcohol, Smoking and Substance Involvement Screening Test (ASSIST)
frequency of use (0 = none, 1 = monthly, 2 = weekly, 3 = daily, 4 = multiple times daily); "cudit_score" = Cannabis Use Disorders Identification Test (CUDIT) total score;
"cigarette_assist" = ASSIST frequency of combustible cigarette use; "ecigarette_use" = ASSIST frequency of electronic cigarette use.}
\format{
An object of class \code{data.frame} with 500 rows and 10 columns.
}
\usage{
data(reheat_data)
}
\description{
An example data set from simulated data. The variables are as follows: "id" = participant unique identifier; "group" = group participant belongs to (A or B);
"audit_score" = Alcohol Use Disorders Identification Test (AUDIT) total score; "sip_score" = Short Inventory of Problems (SIP) total score;
"drinks_per_week" = total number of standard drinks endorsed in a typical week; "cannabis_assist" = Alcohol, Smoking and Substance Involvement Screening Test (ASSIST)
frequency of use (0 = none, 1 = monthly, 2 = weekly, 3 = daily, 4 = multiple times daily); "cudit_score" = Cannabis Use Disorders Identification Test (CUDIT) total score;
"cigarette_assist" = ASSIST frequency of combustible cigarette use; "ecigarette_use" = ASSIST frequency of electronic cigarette use.
}
\examples{
\dontrun{
data("reheat_data")
}
}
\keyword{datasets}
